refactor: rename wallet interface file
